rename render.js to create-element.js
