+++ lib/patch-op.js
@@ -1,4 +1,4 @@
-var render = require("../render")
+var render = require("../create-element")
 var isWidget = require("./is-widget")
 var isString = require("./is-string")
 var isVNode = require("./is-virtual-dom")
